refactor: add deleteEmployee ResponseStatusException

+++ src/main/java/com/example/demo/service/EmployeeService.java
@@ -38,14 +38,17 @@ public class EmployeeService {
 
     public Employee updateEmployee(@PathVariable int id, @RequestBody Employee updatedEmployee) {
         Employee employee = employeeRepository.getEmployeeById(id);
-
         if (employee == null) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Employee not found with id: " + id);
         }
         return employeeRepository.updateEmployee(id, updatedEmployee);
     }
 
-    public void deleteEmployee(@PathVariable int id) {
+    public void deleteEmployee(int id) {
+        Employee employee = employeeRepository.getEmployeeById(id);
+        if (employee == null) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Employee not found with id: " + id);
+        }
         employeeRepository.deleteEmployee(id);
     }
